# api/openapi.yaml
openapi: 3.0.3
info:
  title: Incident Management API
  description: |
    REST API for managing IT incidents, owners, and optional GitHub repository integration.
    Provides full CRUD operations along with filtering, searching, and statistics endpoints.
  version: 1.0.0
  contact:
    name: API Support
    email: support@incident-now.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.incident-now.io/v1
    description: Production server
...
paths:
...
  /owners:
    get:
      tags:
        - Owners
      summary: List all owners
      description: Retrieve a paginated list of owners
      operationId: listOwners
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: team
          in: query
          description: Filter by team name
          schema:
            type: string
        - name: role
          in: query
          description: Filter by role
          schema:
            $ref: '#/components/schemas/OwnerRole'
        - name: search
          in: query
          description: Search in name and email
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Owners
      summary: Create a new owner
      description: Create a new owner/user
      operationId: createOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerCreate'
      responses:
        '201':
          description: Owner created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Owner with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /owners/{ownerId}:
    get:
      tags:
        - Owners
      summary: Get owner by ID
      description: Retrieve a specific owner by their ID
      operationId: getOwnerById
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Owners
      summary: Update an owner
      description: Update an existing owner (full replacement)
      operationId: updateOwner
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerUpdate'
      responses:
        '200':
          description: Owner updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Owner with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Owners
      summary: Partially update an owner
      description: Update specific fields of an existing owner
      operationId: patchOwner
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerPatch'
      responses:
        '200':
          description: Owner updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Owner with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Owners
      summary: Delete an owner
      description: Delete an owner by their ID (soft delete - sets active to false)
      operationId: deleteOwner
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
      responses:
        '204':
          description: Owner deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete owner with active incidents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /owners/{ownerId}/incidents:
    get:
      tags:
        - Owners
      summary: Get incidents owned by user
      description: Retrieve all incidents where this owner is the primary owner
      operationId: getOwnerIncidents
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter by incident status
          schema:
            $ref: '#/components/schemas/IncidentStatus'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /owners/{ownerId}/assigned-incidents:
    get:
      tags:
        - Owners
      summary: Get incidents assigned to user
      description: Retrieve all incidents where this owner is assigned
      operationId: getOwnerAssignedIncidents
      parameters:
        - $ref: '#/components/parameters/OwnerIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter by incident status
          schema:
            $ref: '#/components/schemas/IncidentStatus'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== STATISTICS ====================
  /statistics/summary:
    get:
      tags:
        - Statistics
      summary: Get incident summary statistics
      description: Retrieve summary statistics for incidents
      operationId: getIncidentSummary
      parameters:
        - name: from
          in: query
          description: Start date for statistics
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date for statistics
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/by-status:
    get:
      tags:
        - Statistics
      summary: Get incidents grouped by status
      description: Retrieve incident counts grouped by status
      operationId: getIncidentsByStatus
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/by-priority:
    get:
      tags:
        - Statistics
      summary: Get incidents grouped by priority
      description: Retrieve incident counts grouped by priority
      operationId: getIncidentsByPriority
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriorityCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/by-category:
    get:
      tags:
        - Statistics
      summary: Get incidents grouped by category
      description: Retrieve incident counts grouped by category
      operationId: getIncidentsByCategory
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/by-owner:
    get:
      tags:
        - Statistics
      summary: Get incidents grouped by owner
      description: Retrieve incident counts grouped by owner
      operationId: getIncidentsByOwner
      parameters:
        - name: limit
          in: query
          description: Limit number of owners returned
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnerCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/resolution-time:
    get:
      tags:
        - Statistics
      summary: Get average resolution time
      description: Retrieve average resolution time statistics
      operationId: getResolutionTimeStats
      parameters:
        - name: from
          in: query
          description: Start date for statistics
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date for statistics
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          description: Group results by time period
          schema:
            type: string
            enum: [day, week, month]
            default: week
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolutionTimeStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /statistics/trends:
    get:
      tags:
        - Statistics
      summary: Get incident trends
      description: Retrieve incident creation trends over time
      operationId: getIncidentTrends
      parameters:
        - name: from
          in: query
          description: Start date for trends
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date for trends
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          description: Group results by time period
          schema:
            type: string
            enum: [day, week, month]
            default: week
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrendData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  schemas:
...
    IncidentStatus:
      type: string
      enum:
        - open
        - in_progress
        - on_hold
        - resolved
        - closed
      description: Current status of the incident
    Priority:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
      description: Priority level of the incident
...
    IncidentCategory:
      type: string
      enum:
        - network
        - hardware
        - software
        - security
        - database
        - cloud_infrastructure
        - application
        - performance
        - access_permissions
        - other
      description: Category of the incident
    OwnerRole:
      type: string
      enum:
        - engineer
        - senior_engineer
        - team_lead
        - manager
        - admin
      description: Role of the owner/user

    # ----- GitHub Repository -----
...
    IncidentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ----- Owner -----
    Owner:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the owner
          readOnly: true
        name:
          type: string
          description: Full name of the owner
          maxLength: 255
        email:
          type: string
          format: email
          description: Email address
        phone:
          type: string
          description: Phone number
          nullable: true
        avatarUrl:
          type: string
          format: uri
          description: URL to profile picture
          nullable: true
        team:
          type: string
          description: Team name
        role:
          $ref: '#/components/schemas/OwnerRole'
        department:
          type: string
          description: Department name
          nullable: true
        timezone:
          type: string
          description: Timezone (IANA format)
          example: "America/New_York"
        slackHandle:
          type: string
          description: Slack username/handle
          nullable: true
        githubUsername:
          type: string
          description: GitHub username
          nullable: true
        active:
          type: boolean
          description: Whether the owner is active
        onCall:
          type: boolean
          description: Whether the owner is currently on-call
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the owner was created
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the owner was last updated
          readOnly: true
      required:
        - id
        - name
        - email
        - team
        - role
        - active
        - createdAt
        - updatedAt
    OwnerSummary:
      type: object
      description: Condensed owner information for embedding in other entities
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
          nullable: true
        team:
          type: string
        role:
          $ref: '#/components/schemas/OwnerRole'
    OwnerCreate:
      type: object
      required:
        - name
        - email
        - team
        - role
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
        avatarUrl:
          type: string
          format: uri
        team:
          type: string
        role:
          $ref: '#/components/schemas/OwnerRole'
        department:
          type: string
        timezone:
          type: string
          example: "America/New_York"
        slackHandle:
          type: string
        githubUsername:
          type: string
        onCall:
          type: boolean
          default: false
    OwnerUpdate:
      type: object
      required:
        - name
        - email
        - team
        - role
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
        avatarUrl:
          type: string
          format: uri
        team:
          type: string
        role:
          $ref: '#/components/schemas/OwnerRole'
        department:
          type: string
        timezone:
          type: string
        slackHandle:
          type: string
        githubUsername:
          type: string
        active:
          type: boolean
        onCall:
          type: boolean
    OwnerPatch:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
        avatarUrl:
          type: string
          format: uri
        team:
          type: string
        role:
          $ref: '#/components/schemas/OwnerRole'
        department:
          type: string
        timezone:
          type: string
        slackHandle:
          type: string
        githubUsername:
          type: string
        active:
          type: boolean
        onCall:
          type: boolean
    OwnerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Owner'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ----- Comments -----
...
    IncidentSummary:
      type: object
      properties:
        totalIncidents:
          type: integer
        openIncidents:
          type: integer
        inProgressIncidents:
          type: integer
        onHoldIncidents:
          type: integer
        resolvedIncidents:
          type: integer
        closedIncidents:
          type: integer
        criticalIncidents:
          type: integer
        averageResolutionTimeMinutes:
          type: number
          format: float
        averageTimeToAcknowledgeMinutes:
          type: number
          format: float
        slaBreachCount:
          type: integer
        slaCompliancePercentage:
          type: number
          format: float
    StatusCount:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/IncidentStatus'
        count:
          type: integer
    PriorityCount:
      type: object
      properties:
        priority:
          $ref: '#/components/schemas/Priority'
        count:
          type: integer
    CategoryCount:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/IncidentCategory'
        count:
          type: integer
    OwnerCount:
      type: object
      properties:
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        totalIncidents:
          type: integer
        openIncidents:
          type: integer
        resolvedIncidents:
          type: integer
    ResolutionTimeStats:
      type: object
      properties:
        averageMinutes:
          type: number
          format: float
        medianMinutes:
          type: number
          format: float
        minMinutes:
          type: integer
        maxMinutes:
          type: integer
        percentile90Minutes:
          type: number
          format: float
        data:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              averageMinutes:
                type: number
                format: float
              count:
                type: integer
    TrendData:
      type: object
      properties:
        period:
          type: string
          description: Time period (e.g., "2026-02-01" or "2026-W07")
        created:
          type: integer
          description: Number of incidents created
        resolved:
          type: integer
          description: Number of incidents resolved
        closed:
          type: integer
          description: Number of incidents closed

    # ----- Common -----
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        pageSize:
          type: integer
          description: Number of items per page
        totalItems:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Detailed validation errors
        traceId:
          type: string
          description: Trace ID for debugging
        timestamp:
          type: string
          format: date-time

  # ==================== PARAMETERS ====================
  parameters:
...
    OwnerIdParam:
      name: ownerId
      in: path
      required: true
      description: Unique identifier of the owner
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable entity - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ==================== SECURITY ====================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
security:
  - bearerAuth: []
  - apiKeyAuth: []